Move LinodeClient code to services; use controllerutil.RemoveFinalizer as other controllers; remove unnecessary check
